Updated/Refactored Libmemcached tests
