+++ tests/unit/Cache/Backend/Libmemcached/GetCest.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 
 namespace Phalcon\Test\Unit\Cache\Backend\Libmemcached;
 
+use Phalcon\Test\Fixtures\Traits\Cache\LibmemcachedTrait;
 use UnitTester;
 
 /**
@@ -19,6 +20,8 @@ use UnitTester;
  */
 class GetCest
 {
+    use LibmemcachedTrait;
+
     /**
      * Tests Phalcon\Cache\Backend\Libmemcached :: get()
      *
@@ -30,6 +33,81 @@ class GetCest
     public function cacheBackendLibmemcachedGet(UnitTester $I)
     {
         $I->wantToTest('Cache\Backend\Libmemcached - get()');
-        $I->skipTest('Need implementation');
+
+        $key  = 'data-get';
+        $data = [uniqid(), gethostname(), microtime(), get_include_path(), time()];
+
+        $cache = $this->getDataCache();
+
+        $I->haveInLibmemcached($key, serialize($data));
+        $I->assertEquals($data, $cache->get('data-get'));
+
+        $I->assertNull($cache->get('non-existent-key'));
+
+        $data = 'sure, nothing interesting';
+
+        $I->haveInLibmemcached($key, serialize($data));
+        $I->assertEquals($data, $cache->get('data-get'));
+
+        $I->assertNull($cache->get('non-existent-key-2'));
+    }
+
+    /**
+     * @issue https://github.com/phalcon/cphalcon/issue/13947
+     *
+     * @param UnitTester $I
+     *
+     * @author Cameron Hall <me@chall.id.au>
+     * @since  2018-01-05
+     */
+    public function shouldGetEmptyValuesInsteadOfNull(UnitTester $I)
+    {
+        $I->wantToTest('Issue #13497');
+
+        $cache = $this->getCache(new None());
+
+        $cache->save('empty.array', []);
+        $I->assertEquals([], $cache->get('empty.array'));
+
+        $cache->save('empty.int', 0);
+        $I->assertEquals(0, $cache->get('empty.int'));
+
+        $cache->save('empty.string', '');
+        $I->assertEquals('', $cache->get('empty.string'));
+
+        $cache->save('empty.bool', false);
+        $I->assertFalse($cache->get('empty.bool'));
+
+        $I->assertNull(null, $cache->get('valuethatdoesnotexist'));
+
+        $cache->save('empty.null', null);
+        $I->assertNull($cache->get('null'));
+    }
+
+    /**
+     * @param UnitTester $I
+     * @issue https://github.com/phalcon/cphalcon/issues/13092
+     */
+    public function shouldGetTheSameValueRegardlessOfTheNumberOfRequests(UnitTester $I)
+    {
+        $I->wantTo('Get the same data from the Memcache regardless of the number of requests');
+
+        $key  = 'libmemcached-data-get-test';
+        $data = 'this is a test';
+
+        $cache = $this->getDataCache();
+
+        $I->haveInLibmemcached($key, serialize($data));
+
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
+        $I->assertEquals($data, $cache->get($key));
     }
 }
