+++ tests/unit/Cache/Backend/Libmemcached/IncrementCest.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 
 namespace Phalcon\Test\Unit\Cache\Backend\Libmemcached;
 
+use Phalcon\Test\Fixtures\Traits\Cache\LibmemcachedTrait;
 use UnitTester;
 
 /**
@@ -19,6 +20,8 @@ use UnitTester;
  */
 class IncrementCest
 {
+    use LibmemcachedTrait;
+
     /**
      * Tests Phalcon\Cache\Backend\Libmemcached :: increment()
      *
@@ -29,7 +32,21 @@ class IncrementCest
      */
     public function cacheBackendLibmemcachedIncrement(UnitTester $I)
     {
-        $I->wantToTest('Cache\Backend\Libmemcached - increment()');
-        $I->skipTest('Need implementation');
+        $I->wantTo('Cache\Backend\Libmemcached - increment()');
+
+        $key   = 'increment';
+        $cache = $this->getDataCache();
+
+        $I->dontSeeInLibmemcached($key);
+        $I->haveInLibmemcached($key, 1);
+
+        $I->assertEquals(2, $cache->increment($key));
+        $I->seeInLibmemcached($key, 2);
+
+        $I->assertEquals(4, $cache->increment($key, 2));
+        $I->seeInLibmemcached($key, 4);
+
+        $I->assertEquals(14, $cache->increment($key, 10));
+        $I->seeInLibmemcached($key, 14);
     }
 }
