+++ tests/unit/Cache/Backend/Libmemcached/QueryKeysCest.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 
 namespace Phalcon\Test\Unit\Cache\Backend\Libmemcached;
 
+use Phalcon\Test\Fixtures\Traits\Cache\LibmemcachedTrait;
 use UnitTester;
 
 /**
@@ -19,6 +20,8 @@ use UnitTester;
  */
 class QueryKeysCest
 {
+    use LibmemcachedTrait;
+
     /**
      * Tests Phalcon\Cache\Backend\Libmemcached :: queryKeys()
      *
@@ -30,6 +33,73 @@ class QueryKeysCest
     public function cacheBackendLibmemcachedQueryKeys(UnitTester $I)
     {
         $I->wantToTest('Cache\Backend\Libmemcached - queryKeys()');
-        $I->skipTest('Need implementation');
+
+        $lifetime = 20;
+        $statsKey = '_PHCM';
+        $cache    = $this->getDataCache($statsKey);
+
+        $I->haveInLibmemcached("a", 1);
+        $I->haveInLibmemcached("b", 2);
+        $I->haveInLibmemcached("c", 3);
+
+        $I->haveInLibmemcached($statsKey, ['a' => $lifetime, 'b' => $lifetime, 'c' => $lifetime]);
+
+        $keys = $cache->queryKeys();
+        sort($keys);
+
+        $I->assertEquals(['a', 'b', 'c'], $keys);
     }
+
+    public function emptyQueryKeys(UnitTester $I)
+    {
+        $I->wantToTest('Getting empty keys list by using Libmemcached as cache backend');
+
+        $lifetime = 20;
+        $statsKey = '_PHCM';
+        $cache    = $this->getDataCache($statsKey);
+
+        $I->assertEquals([], $cache->queryKeys());
+    }
+    
+    /**
+     * @issue https://github.com/phalcon/cphalcon/issues/11024
+     * @param UnitTester $I
+     */
+    public function prefixedQueryKeys(UnitTester $I)
+    {
+        $I->wantToTest('Getting prefixed cache keys by using Libmemcached as cache backend');
+
+        $lifetime = 20;
+        $statsKey = '_PHCM';
+        $cache    = $this->getDataCache($statsKey);
+
+        $I->haveInLibmemcached('prefix1-myKey', ['a', 'b']);
+        $I->haveInLibmemcached('prefix2-myKey', ['x', 'z']);
+
+        $I->haveInLibmemcached($statsKey, ['prefix1-myKey' => $lifetime, 'prefix2-myKey' => $lifetime]);
+
+        $I->assertEquals([0 => 'prefix1-myKey'], $cache->queryKeys('prefix1'));
+        $I->assertEquals([1 => 'prefix2-myKey'], $cache->queryKeys('prefix2'));
+        $I->assertCount(2, $cache->queryKeys('prefix'));
+
+        $I->assertEquals([], $cache->queryKeys('prefix123'));
+    }
+
+    public function queryKeysWithoutStatsKey(UnitTester $I)
+    {
+        $I->wantTo(
+            'Catch exception during the attempt getting cache keys by using ' .
+            'Libmemcached as cache backend without statsKey'
+        );
+
+        $cache = $this->getDataCache(null, 1);
+
+        $I->expectException(
+            new Exception("Cached keys need to be enabled to use this function (options['statsKey'] == '_PHCM')!"),
+            function () use ($cache) {
+                $cache->queryKeys();
+            }
+        );
+    }
+
 }
