+++ tests/unit/Cache/Backend/LibmemcachedCest.php
@@ -2,13 +2,13 @@
 
 namespace Phalcon\Test\Unit\Cache\Backend;
 
+use Phalcon\Test\Fixtures\Traits\Cache\LibmemcachedTrait;
 use Phalcon\Cache\Backend\Libmemcached;
 use Phalcon\Cache\Exception;
 use Phalcon\Cache\Frontend\Data;
 use Phalcon\Cache\Frontend\Output;
 use Phalcon\Cache\Frontend\None;
 use Phalcon\Cache\Frontend;
-use Phalcon\Cache\FrontendInterface;
 use UnitTester;
 
 /**
@@ -30,245 +30,7 @@ use UnitTester;
  */
 class LibmemcachedCest
 {
-    public function _before(UnitTester $I)
-    {
-        $I->checkExtensionIsLoaded('memcached');
-    }
-
-    public function increment(UnitTester $I)
-    {
-        $I->wantTo('Increment counter by using Libmemcached as cache backend');
-
-        $key   = 'increment';
-        $cache = $this->getDataCache();
-
-        $I->dontSeeInLibmemcached($key);
-        $I->haveInLibmemcached($key, 1);
-
-        $I->assertEquals(2, $cache->increment($key));
-        $I->seeInLibmemcached($key, 2);
-
-        $I->assertEquals(4, $cache->increment($key, 2));
-        $I->seeInLibmemcached($key, 4);
-
-        $I->assertEquals(14, $cache->increment($key, 10));
-        $I->seeInLibmemcached($key, 14);
-    }
-
-    protected function getDataCache($statsKey = null, $ttl = 20)
-    {
-        $config = [];
-        if ($statsKey !== null) {
-            $config['statsKey'] = $statsKey;
-        }
-
-        return $this->getCache(new Data(['lifetime' => $ttl]), $config);
-    }
-
-    public function decrement(UnitTester $I)
-    {
-        $I->wantTo('Decrement counter by using Libmemcached as cache backend');
-
-        $key   = 'decrement';
-        $cache = $this->getDataCache();
-
-        $I->dontSeeInLibmemcached($key);
-        $I->haveInLibmemcached($key, 100);
-
-        $I->assertEquals(99, $cache->decrement($key));
-        $I->seeInLibmemcached($key, 99);
-
-        $I->assertEquals(97, $cache->decrement($key, 2));
-        $I->seeInLibmemcached($key, 97);
-
-        $I->assertEquals(87, $cache->decrement($key, 10));
-        $I->seeInLibmemcached($key, 87);
-    }
-
-    public function get(UnitTester $I)
-    {
-        $I->wantTo('Get data by using Libmemcached as cache backend');
-
-        $key  = 'data-get';
-        $data = [uniqid(), gethostname(), microtime(), get_include_path(), time()];
-
-        $cache = $this->getDataCache();
-
-        $I->haveInLibmemcached($key, serialize($data));
-        $I->assertEquals($data, $cache->get('data-get'));
-
-        $I->assertNull($cache->get('non-existent-key'));
-
-        $data = 'sure, nothing interesting';
-
-        $I->haveInLibmemcached($key, serialize($data));
-        $I->assertEquals($data, $cache->get('data-get'));
-
-        $I->assertNull($cache->get('non-existent-key-2'));
-    }
-
-    /**
-     * @param UnitTester $I
-     * @issue https://github.com/phalcon/cphalcon/issues/13092
-     */
-    public function shouldGetTheSameValueRegardlessOfTheNumberOfRequests(UnitTester $I)
-    {
-        $I->wantTo('Get the same data from the Memcache regardless of the number of requests');
-
-        $key  = 'libmemcached-data-get-test';
-        $data = 'this is a test';
-
-        $cache = $this->getDataCache();
-
-        $I->haveInLibmemcached($key, serialize($data));
-
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-        $I->assertEquals($data, $cache->get($key));
-    }
-
-    public function save(UnitTester $I)
-    {
-        $I->wantTo('Save data by using Libmemcached as cache backend');
-
-        $key  = 'data-save';
-        $data = [uniqid(), gethostname(), microtime(), get_include_path(), time()];
-
-        $cache = $this->getDataCache();
-
-        $I->dontSeeInLibmemcached($key);
-        $cache->save('data-save', $data);
-
-        $I->seeInLibmemcached($key, serialize($data));
-
-        $data = 'sure, nothing interesting';
-
-        $I->dontSeeInLibmemcached('non-existent-key');
-
-        $cache->save('data-save', $data);
-        $I->seeInLibmemcached($key, serialize($data));
-    }
-
-    public function delete(UnitTester $I)
-    {
-        $I->wantTo(/** @lang text */
-            'Delete from cache by using Libmemcached as cache backend'
-        );
-
-        $cache = $this->getDataCache();
-
-        $I->assertFalse($cache->delete('non-existent-keys'));
-
-        $I->haveInLibmemcached('some-key-to-delete', 1);
-
-        $I->assertTrue($cache->delete('some-key-to-delete'));
-        $I->dontSeeInLibmemcached('some-key-to-delete');
-    }
-
-    public function flush(UnitTester $I)
-    {
-        $I->wantTo('Flush cache by using Libmemcached as cache backend');
-
-        $lifetime = 20;
-        $statsKey = '_PHCM';
-        $cache    = $this->getDataCache($statsKey);
-
-        $I->haveInLibmemcached('data-flush-1', 1);
-        $I->haveInLibmemcached('data-flush-2', 2);
-        $I->haveInLibmemcached('data-flush-3', 3);
-
-        $I->haveInLibmemcached(
-            $statsKey,
-            ['data-flush-1' => $lifetime, 'data-flush-2' => $lifetime, 'data-flush-3' => $lifetime]
-        );
-
-        $I->assertTrue($cache->flush());
-
-        $I->dontSeeInLibmemcached('data-flush-1');
-        $I->dontSeeInLibmemcached('data-flush-2');
-        $I->dontSeeInLibmemcached('data-flush-3');
-
-        $I->dontSeeInLibmemcached($statsKey);
-    }
-
-    public function emptyQueryKeys(UnitTester $I)
-    {
-        $I->wantToTest('Getting empty keys list by using Libmemcached as cache backend');
-
-        $lifetime = 20;
-        $statsKey = '_PHCM';
-        $cache    = $this->getDataCache($statsKey);
-
-        $I->assertEquals([], $cache->queryKeys());
-    }
-
-    public function queryKeys(UnitTester $I)
-    {
-        $I->wantToTest('Getting cache keys by using Libmemcached as cache backend');
-
-        $lifetime = 20;
-        $statsKey = '_PHCM';
-        $cache    = $this->getDataCache($statsKey);
-
-        $I->haveInLibmemcached("a", 1);
-        $I->haveInLibmemcached("b", 2);
-        $I->haveInLibmemcached("c", 3);
-
-        $I->haveInLibmemcached($statsKey, ['a' => $lifetime, 'b' => $lifetime, 'c' => $lifetime]);
-
-        $keys = $cache->queryKeys();
-        sort($keys);
-
-        $I->assertEquals(['a', 'b', 'c'], $keys);
-    }
-
-    /**
-     * @issue https://github.com/phalcon/cphalcon/issues/11024
-     * @param UnitTester $I
-     */
-    public function prefixedQueryKeys(UnitTester $I)
-    {
-        $I->wantToTest('Getting prefixed cache keys by using Libmemcached as cache backend');
-
-        $lifetime = 20;
-        $statsKey = '_PHCM';
-        $cache    = $this->getDataCache($statsKey);
-
-        $I->haveInLibmemcached('prefix1-myKey', ['a', 'b']);
-        $I->haveInLibmemcached('prefix2-myKey', ['x', 'z']);
-
-        $I->haveInLibmemcached($statsKey, ['prefix1-myKey' => $lifetime, 'prefix2-myKey' => $lifetime]);
-
-        $I->assertEquals([0 => 'prefix1-myKey'], $cache->queryKeys('prefix1'));
-        $I->assertEquals([1 => 'prefix2-myKey'], $cache->queryKeys('prefix2'));
-        $I->assertCount(2, $cache->queryKeys('prefix'));
-
-        $I->assertEquals([], $cache->queryKeys('prefix123'));
-    }
-
-    public function queryKeysWithoutStatsKey(UnitTester $I)
-    {
-        $I->wantTo(
-            'Catch exception during the attempt getting cache keys by using ' .
-            'Libmemcached as cache backend without statsKey'
-        );
-
-        $cache = $this->getDataCache(null, 1);
-
-        $I->expectException(
-            new Exception("Cached keys need to be enabled to use this function (options['statsKey'] == '_PHCM')!"),
-            function () use ($cache) {
-                $cache->queryKeys();
-            }
-        );
-    }
+    use LibmemcachedTrait;
 
     public function output(UnitTester $I)
     {
@@ -305,50 +67,5 @@ class LibmemcachedCest
 
         $I->assertNull($content);
         $I->dontSeeInLibmemcached('test-output');
-    }
-
-    protected function getOutputCache($ttl = 0)
-    {
-        return $this->getCache(new Output(['lifetime' => $ttl]));
-    }
-
-    protected function getCache(FrontendInterface $frontend, $config = [])
-    {
-        $config = array_merge($config, [
-            'client'  => [],
-            'servers' => [
-                [
-                    'host'   => env('DATA_MEMCACHED_HOST', '127.0.0.1'),
-                    'port'   => env('DATA_MEMCACHED_PORT', 11211),
-                    'weight' => env('DATA_MEMCACHED_WEIGHT', 0),
-                ],
-            ],
-        ]);
-
-        return new Libmemcached($frontend, $config);
-    }
-
-    public function testIssue13497(UnitTester $I)
-    {
-        $I->wantToTest('Issue #13497');
-
-        $cache = $this->getCache(new None());
-
-        $cache->save('empty.array', []);
-        $I->assertEquals([], $cache->get('empty.array'));
-
-        $cache->save('empty.int', 0);
-        $I->assertEquals(0, $cache->get('empty.int'));
-
-        $cache->save('empty.string', '');
-        $I->assertEquals('', $cache->get('empty.string'));
-
-        $cache->save('empty.bool', false);
-        $I->assertFalse($cache->get('empty.bool'));
-
-        $I->assertNull(null, $cache->get('valuethatdoesnotexist'));
-
-        $cache->save('empty.null', null);
-        $I->assertNull($cache->get('null'));
     }
 }
